requirements.txt
openai~=0.28.0
Flask~=3.1.0
pillow~=11.0.0
python-dotenv~=1.0.0